backend/requirements.txt
faiss-cpu
torch
numpy
Pillow
Flask
flask-cors
